% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.aggregate.r
\name{.datashield.aggregate}
\alias{.datashield.aggregate}
\title{Data aggregation}
\usage{
.datashield.aggregate(opal, expr, async = TRUE)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}

\item{expr}{Expression to evaluate.}

\item{async}{Whether the call should be asynchronous.}
}
\value{
The R command ID to be used to retrieve the command result (when async) or the agregated value.
}
\description{
Aggregates the expression result using the specified aggregation method in the current Datashield session.
This operation is asynchronous and non blocking.
}
\keyword{internal}
