% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.assign.r
\name{.datashield.assign.resource}
\alias{.datashield.assign.resource}
\title{Data assignment}
\usage{
.datashield.assign.resource(opal, symbol, value, async = TRUE)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}

\item{value}{Fully qualified name of a resource in Opal.}

\item{async}{Whether the call should be asynchronous.}
}
\value{
The R command ID if the async flag is TRUE and if the wait flag is FALSE and if Opal version is at least 2.1, NULL otherwise.
}
\description{
Assign a Opal value to a R symbol in the current Datashield session.
This operation is asynchronous and non blocking.
}
\keyword{internal}
