% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.command.r
\name{.datashield.commands}
\alias{.datashield.commands}
\title{List the asynchronous commands}
\usage{
.datashield.commands(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Get the list of asynchronous R commands in the remote Datashield session.
}
\keyword{internal}
