% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.command.r
\name{.datashield.commands_rm}
\alias{.datashield.commands_rm}
\title{Remove all asynchronous commands}
\usage{
.datashield.commands_rm(opal)
}
\arguments{
\item{opal}{Opal object or list of opal objects.}
}
\description{
Remove all asynchronous R commands in the remote Datashield session.
}
\keyword{internal}
