% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.method.r
\name{.datashield.methods}
\alias{.datashield.methods}
\title{List Datashield methods}
\usage{
.datashield.methods(opal, type = "aggregate")
}
\arguments{
\item{opal}{Opal object.}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\description{
Get available Datashield methods of a given type.
}
\keyword{internal}
