% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.method.r
\name{.datashield.profiles}
\alias{.datashield.profiles}
\title{List Datashield profiles}
\usage{
.datashield.profiles(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get available Datashield profiles.
}
\keyword{internal}
