% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.symbol.r
\name{.datashield.rm}
\alias{.datashield.rm}
\title{Remove a R symbol}
\usage{
.datashield.rm(opal, symbol)
}
\arguments{
\item{opal}{Opal object.}

\item{symbol}{Name of the R symbol.}
}
\description{
Remove a symbol from the current Datashield session.
}
\keyword{internal}
