% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.symbol.r
\name{.datashield.symbols}
\alias{.datashield.symbols}
\title{List R symbols}
\usage{
.datashield.symbols(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get the R symbols available after the datashield.assign calls in the current Datashield session.
}
\keyword{internal}
