% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.r
\name{.datashield.workspace_restore}
\alias{.datashield.workspace_restore}
\title{Restore a saved workspace in the current DataSHIELD session.}
\usage{
.datashield.workspace_restore(opal, ws)
}
\arguments{
\item{opal}{Opal object.}

\item{ws}{The workspace name}
}
\description{
Restore a saved workspace in the current DataSHIELD session.
}
\keyword{internal}
