% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.r
\name{.datashield.workspace_rm}
\alias{.datashield.workspace_rm}
\title{Remove a DataSHIELD workspace from a opal.}
\usage{
.datashield.workspace_rm(opal, ws)
}
\arguments{
\item{opal}{Opal object.}

\item{ws}{The workspace name}
}
\description{
Remove a DataSHIELD workspace from a opal.
}
\keyword{internal}
