% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.r
\name{.datashield.workspace_save}
\alias{.datashield.workspace_save}
\title{Save current session in a DataSHIELD workspace.}
\usage{
.datashield.workspace_save(opal, ws)
}
\arguments{
\item{opal}{Opal object.}

\item{ws}{The workspace name}
}
\description{
Save current session in a DataSHIELD workspace.
}
\keyword{internal}
