% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datashield.workspace.r
\name{.datashield.workspaces}
\alias{.datashield.workspaces}
\title{Get the DataSHIELD workspaces.}
\usage{
.datashield.workspaces(opal)
}
\arguments{
\item{opal}{Opal object.}
}
\description{
Get the DataSHIELD workspaces.
}
\keyword{internal}
