% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getDatashieldSession}
\alias{.getDatashieldSession}
\title{Get R session from opal object, create a new Datashield R session if not found.}
\usage{
.getDatashieldSession(opal)
}
\description{
Get R session from opal object, create a new Datashield R session if not found.
}
\keyword{internal}
