% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getDatashieldSessionId}
\alias{.getDatashieldSessionId}
\title{Extract R session Id from opal object, create a new Datashield R session if not found.}
\usage{
.getDatashieldSessionId(opal)
}
\description{
Extract R session Id from opal object, create a new Datashield R session if not found.
}
\keyword{internal}
