% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.newDatashieldSession}
\alias{.newDatashieldSession}
\title{Create a new Datashield R session in Opal.}
\usage{
.newDatashieldSession(opal, restore = NULL)
}
\description{
Create a new Datashield R session in Opal.
}
\keyword{internal}
