% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\docType{methods}
\name{dsAssignExpr,OpalConnection-method}
\alias{dsAssignExpr,OpalConnection-method}
\title{Assign the result of an expression}
\usage{
\S4method{dsAssignExpr}{OpalConnection}(conn, symbol, expr)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} object.}

\item{symbol}{Name of the R symbol.}

\item{expr}{A R expression with allowed assign functions calls.}
}
\value{
A \code{\link{OpalResult-class}} object.
}
\description{
Assign a result of the execution of an expression in the DataSHIELD R session.
}
