% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsDisconnect,OpalConnection-method}
\alias{dsDisconnect,OpalConnection-method}
\title{Disconnect from a Opal server}
\usage{
\S4method{dsDisconnect}{OpalConnection}(conn, save = NULL)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}

\item{save}{Save the DataSHIELD R session with provided ID (must be a character string).}
}
\description{
Disconnect from a Opal server and release all R resources. If a workspace ID is provided, the DataSHIELD
R session will be saved before being destroyed.
}
\examples{
\dontrun{
con <- dsConnect(DSOpal::Opal(), "server1", "username", "password", "https://opal.example.org")
con
dsDisconnect(con)
}

}
