% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalResult.R
\docType{methods}
\name{dsFetch,OpalResult-method}
\alias{dsFetch,OpalResult-method}
\title{Fetch the result}
\usage{
\S4method{dsFetch}{OpalResult}(res)
}
\arguments{
\item{res}{\code{\link{OpalResult-class}} object.}
}
\value{
TRUE if table exists.
}
\description{
Fetch the DataSHIELD operation result.
}
