% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsKeepAlive,OpalConnection-method}
\alias{dsKeepAlive,OpalConnection-method}
\title{Keep connection with a Opal server alive}
\usage{
\S4method{dsKeepAlive}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\description{
Makes a dummy web service request.
}
\examples{
\dontrun{
con <- dsConnect(DSOpal::Opal(), "server1", "username", "password", "https://opal.example.org")
dsKeepAlive(con)
dsDisconnect(con)
}

}
