% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsListMethods,OpalConnection-method}
\alias{dsListMethods,OpalConnection-method}
\title{List methods}
\usage{
\S4method{dsListMethods}{OpalConnection}(conn, type = "aggregate")
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}

\item{type}{Type of the method: "aggregate" (default) or "assign".}
}
\value{
A data frame.
}
\description{
List methods defined in the DataSHIELD configuration.
}
