% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsListPackages,OpalConnection-method}
\alias{dsListPackages,OpalConnection-method}
\title{List packages}
\usage{
\S4method{dsListPackages}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
A data frame with columns: name, version.
}
\description{
List packages defined in the DataSHIELD configuration.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsListPackages(con)
dsDisconnect(con)
}

}
