% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsListResources,OpalConnection-method}
\alias{dsListResources,OpalConnection-method}
\title{List Opal resources}
\usage{
\S4method{dsListResources}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
The fully qualified names of the resources.
}
\description{
List Opal resources that may be accessible for performing DataSHIELD operations.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsListResources(con)
dsDisconnect(con)
}

}
