% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsListTables,OpalConnection-method}
\alias{dsListTables,OpalConnection-method}
\title{List Opal tables}
\usage{
\S4method{dsListTables}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
The fully qualified names of the tables.
}
\description{
List Opal tables that may be accessible for performing DataSHIELD operations.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsListTables(con)
dsDisconnect(con)
}

}
