% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\docType{methods}
\name{dsListWorkspaces,OpalConnection-method}
\alias{dsListWorkspaces,OpalConnection-method}
\title{List workspaces}
\usage{
\S4method{dsListWorkspaces}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
A data frame.
}
\description{
List workspaces saved in the data repository.
}
