% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSOpal-package.R
\docType{package}
\name{DSOpal-package}
\alias{DSOpal}
\alias{DSOpal-package}
\title{DSOpal: 'DataSHIELD' Implementation for 'Opal'}
\description{
'DataSHIELD' is an infrastructure and series of R packages that enables the remote and 'non-disclosive' analysis of sensitive research data. This package is the 'DataSHIELD' interface implementation for 'Opal', which is the data integration application for biobanks by 'OBiBa'. Participant data, once collected from any data source, must be integrated and stored in a central data repository under a uniform model. 'Opal' is such a central repository. It can import, process, validate, query, analyze, report, and export data. 'Opal' is the reference implementation of the 'DataSHIELD' infrastructure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datashield/DSOpal/}
  \item \url{https://datashield.github.io/DSOpal/}
  \item \url{https://www.obiba.org}
  \item \url{https://www.obiba.org/pages/products/opal/}
  \item \url{https://www.datashield.org}
  \item \url{https://academic.oup.com/ije/article/43/6/1929/707730}
  \item \url{https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1008880}
  \item Report bugs at \url{https://github.com/datashield/DSOpal/issues/}
}

}
\author{
\strong{Maintainer}: Yannick Marcon \email{yannick.marcon@obiba.org} (\href{https://orcid.org/0000-0003-0138-2023}{ORCID})

Other contributors:
\itemize{
  \item Becca Wilson (\href{https://orcid.org/0000-0003-2294-593X}{ORCID}) [contributor]
  \item OBiBa group \email{info@obiba.org} [copyright holder]
}

}
\keyword{internal}
