% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsRestoreWorkspace,OpalConnection-method}
\alias{dsRestoreWorkspace,OpalConnection-method}
\title{Restore workspace}
\usage{
\S4method{dsRestoreWorkspace}{OpalConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Restore workspace from the data repository.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsListWorkspaces(con)
dsRestoreWorkspace(con, "foo")
dsDisconnect(con)
}

}
