% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsRmSymbol,OpalConnection-method}
\alias{dsRmSymbol,OpalConnection-method}
\title{Remove a R symbol}
\usage{
\S4method{dsRmSymbol}{OpalConnection}(conn, symbol)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}

\item{symbol}{Name of the R symbol.}
}
\description{
Remoe a symbol living in the DataSHIELD R session.
}
