% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsRmWorkspace,OpalConnection-method}
\alias{dsRmWorkspace,OpalConnection-method}
\title{Remove a workspace}
\usage{
\S4method{dsRmWorkspace}{OpalConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Remove a workspace on the data repository.
}
