% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\docType{methods}
\name{dsSaveWorkspace,OpalConnection-method}
\alias{dsSaveWorkspace,OpalConnection-method}
\title{Save workspace}
\usage{
\S4method{dsSaveWorkspace}{OpalConnection}(conn, name)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}

\item{name}{Name of the workspace.}
}
\description{
Save workspace on the data repository.
}
