% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logindata.R
\docType{data}
\name{logindata.opal.demo}
\alias{logindata.opal.demo}
\title{DataSHIELD login data file}
\description{
DataSHIELD login data file based on Opal demo server, with simulated data. Note that
the Opal demo server is rebuilt every day and is possibly not accessible.
}
\references{
\url{https://opal-demo.obiba.org}
}
\keyword{data}
