% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\docType{class}
\name{OpalConnection-class}
\alias{OpalConnection-class}
\title{Class OpalConnection.}
\description{
An Opal connection implementing the DataSHIELD Interface (DSI) \code{\link{DSConnection-class}}.
}
\keyword{internal}
