% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalDriver.R
\docType{class}
\name{OpalDriver-class}
\alias{OpalDriver-class}
\title{Class OpalDriver with constructor Opal.}
\description{
An Opal driver implementing the DataSHIELD Interface (DSI) \code{\link{DSDriver-class}}.
This class should always be initialized with the \code{\link{Opal}} function.
It returns a singleton that allows you to connect to Opal.
}
\keyword{internal}
