% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalResult.R
\docType{class}
\name{OpalResult-class}
\alias{OpalResult-class}
\title{Class OpalResult.}
\description{
An Opal result implementing the DataSHIELD Interface (DSI)  \code{\link{DSResult-class}}.
}
\keyword{internal}
