% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalSession.R
\docType{class}
\name{OpalSession-class}
\alias{OpalSession-class}
\title{Class OpalSession.}
\description{
An Opal session implementing the DataSHIELD Interface (DSI)  \code{\link[DSI]{DSSession-class}}.
}
\keyword{internal}
