% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalResult.R
\name{dsFetch,OpalResult-method}
\alias{dsFetch,OpalResult-method}
\title{Fetch the result}
\usage{
\S4method{dsFetch}{OpalResult}(res)
}
\arguments{
\item{res}{\code{\link{OpalResult-class}} object.}
}
\value{
TRUE if table exists.
}
\description{
Fetch the DataSHIELD operation result.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsAssignExpr(con, "C", as.symbol("c(1, 2, 3)"))
res <- dsAggregate(con, as.symbol("length(C)"))
length <- dsFetch(res)
dsDisconnect(con)
}

}
