% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsIsAsync,OpalConnection-method}
\alias{dsIsAsync,OpalConnection-method}
\title{Opal asynchronous support}
\usage{
\S4method{dsIsAsync}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
The named list of logicals detailing the asynchronicity support.
}
\description{
List that Opal supports asynchronicity on all DataSHIELD operations.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsIsAsync(con)
dsDisconnect(con)
}

}
