% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsListProfiles,OpalConnection-method}
\alias{dsListProfiles,OpalConnection-method}
\title{List profiles}
\usage{
\S4method{dsListProfiles}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
A list containing the "available" character vector of profile names and the "current" profile (in case a default one was assigned).
}
\description{
List profiles defined in the DataSHIELD configuration.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsListProfiles(con)
dsDisconnect(con)
}

}
