% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsListSymbols,OpalConnection-method}
\alias{dsListSymbols,OpalConnection-method}
\title{List R symbols}
\usage{
\S4method{dsListSymbols}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
A character vector.
}
\description{
List symbols living in the DataSHIELD R session.
}
