% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpalConnection.R
\name{dsListWorkspaces,OpalConnection-method}
\alias{dsListWorkspaces,OpalConnection-method}
\title{List workspaces}
\usage{
\S4method{dsListWorkspaces}{OpalConnection}(conn)
}
\arguments{
\item{conn}{\code{\link{OpalConnection-class}} class object}
}
\value{
A data frame with columns: name, lastAccessDate, size.
}
\description{
List workspaces saved in the data repository.
}
\examples{
\dontrun{
con <- dbConnect(DSOpal::Opal(), "server1",
  "username", "password", "https://opal.example.org")
dsListWorkspaces(con)
dsDisconnect(con)
}

}
